execution/requirements.txt
# Atliso Chatbot Execution Layer Dependencies

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# OpenAI
openai>=1.3.0

# Google Calendar
google-auth>=2.23.0
google-api-python-client>=2.108.0

# Email
aiosmtplib>=3.0.0

# PDF processing
PyPDF2>=3.0.0
httpx>=0.25.0

# Utilities
pydantic>=2.5.0
python-dateutil>=2.8.2
beautifulsoup4>=4.12.0
